# mini-study-backend/internal/docs/swagger.yaml
definitions:
  dto.AdminCreateBannerRequest:
    properties:
      end_at:
        type: string
      image_url:
        type: string
      link_url:
        type: string
      sort_order:
        type: integer
      start_at:
        type: string
      status:
        type: boolean
      title:
        maxLength: 255
        minLength: 1
        type: string
      visible_roles:
        enum:
        - employee
        - manager
        - both
        type: string
    required:
    - image_url
    - link_url
    - title
    type: object
  dto.AdminCreateContentRequest:
    properties:
      category_id:
        type: integer
      cover_url:
        type: string
      duration_seconds:
        type: integer
      file_path:
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      summary:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - doc
        - video
        type: string
      visible_roles:
        enum:
        - employee
        - manager
        - both
        type: string
    required:
    - category_id
    - file_path
    - title
    - type
    type: object
...
  dto.AdminUpdateBannerRequest:
    properties:
      end_at:
        type: string
      image_url:
        type: string
      link_url:
        type: string
      sort_order:
        type: integer
      start_at:
        type: string
      status:
        type: boolean
      title:
        maxLength: 255
        minLength: 1
        type: string
      visible_roles:
        enum:
        - employee
        - manager
        - both
        type: string
    type: object
  dto.AdminUpdateContentRequest:
    properties:
      category_id:
        type: integer
      cover_url:
        type: string
      duration_seconds:
        type: integer
      file_path:
        type: string
      status:
        enum:
        - draft
        - published
        - offline
        type: string
      summary:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - doc
        - video
        type: string
      visible_roles:
        enum:
        - employee
        - manager
        - both
        type: string
    type: object
...
  dto.BannerResponse:
    properties:
      end_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      link_url:
        type: string
      sort_order:
        type: integer
      start_at:
        type: string
      status:
        type: boolean
      title:
        type: string
      visible_roles:
        type: string
    type: object
  dto.ContentCategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      role_scope:
        type: string
      sort_order:
        type: integer
    type: object
  dto.ContentResponse:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      cover_url:
        type: string
      duration_seconds:
        type: integer
      file_path:
        type: string
      id:
        type: integer
      publish_at:
        type: string
      status:
        type: string
      summary:
        type: string
      title:
        type: string
      type:
        type: string
      visible_roles:
        type: string
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/admin/banners:
    get:
      description: 管理员可按状态筛选轮播图
      parameters:
      - description: 是否启用 true/false
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BannerResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 管理员查询轮播图列表
      tags:
      - 管理后台-轮播
    post:
      consumes:
      - application/json
      description: 管理员上传图片并配置跳转链接、可见角色等信息
      parameters:
      - description: 轮播图信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BannerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 管理员创建轮播图
      tags:
      - 管理后台-轮播
  /api/v1/admin/banners/{id}:
    put:
      consumes:
      - application/json
      description: 管理员可修改轮播图基础信息、状态及可见范围
      parameters:
      - description: 轮播图ID
        in: path
        name: id
        required: true
        type: integer
      - description: 轮播图信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BannerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 管理员更新轮播图
      tags:
      - 管理后台-轮播
  /api/v1/admin/contents:
    get:
      description: 管理员可按分类、类型与状态筛选内容
      parameters:
      - description: 分类ID
        in: query
        name: category_id
        type: integer
      - description: 内容类型(doc/video)
        in: query
        name: type
        type: string
      - description: 内容状态(draft/published/offline)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ContentResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 管理员查询内容列表
      tags:
      - 管理后台-内容
    post:
      consumes:
      - application/json
      description: 管理员上传文件并设置分类、可见角色后创建内容
      parameters:
      - description: 内容信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ContentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 管理员创建内容
      tags:
      - 管理后台-内容
  /api/v1/admin/contents/{id}:
    put:
      consumes:
      - application/json
      description: 管理员可修改内容基础信息、状态与可见范围
      parameters:
      - description: 内容ID
        in: path
        name: id
        required: true
        type: integer
      - description: 内容信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ContentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 管理员更新内容
      tags:
      - 管理后台-内容
...
  /api/v1/banners:
    get:
      description: 返回当前登录用户基于角色可见的轮播图列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BannerResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 查询可见轮播图
      tags:
      - 轮播
  /api/v1/contents:
    get:
      description: 根据分类与类型筛选当前用户可访问的已发布内容
      parameters:
      - description: 分类ID
        in: query
        name: category_id
        type: integer
      - description: 内容类型(doc/video)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ContentResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 查询已发布内容
      tags:
      - 内容
  /api/v1/contents/{id}:
    get:
      description: 获取指定已发布内容的详细信息
      parameters:
      - description: 内容ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ContentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 查询内容详情
      tags:
      - 内容
  /api/v1/contents/categories:
    get:
      description: 返回当前登录用户基于角色可访问的内容分类列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ContentCategoryResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 查询可见内容分类
      tags:
      - 内容
...
swagger: "2.0"